requirements.txt
nemoguardrails 
langchain_openai
guardrails-ai
python-dotenv